# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FalconTrust RAG Candidate API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: rag
    description: RAG candidate pipeline
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /rag/status:
    get:
      operationId: getRagStatus
      tags: [rag]
      summary: Get ingestion and index status
      responses:
        "200":
          description: Current RAG status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagStatus"

  /rag/documents:
    get:
      operationId: listDocuments
      tags: [rag]
      summary: List all ingested documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RagDocument"
    delete:
      operationId: deleteDocumentBySelector
      tags: [rag]
      summary: Remove one ingested document from the candidate index
      parameters:
        - in: query
          name: documentId
          schema:
            type: string
          description: Explicit document ID to remove
        - in: query
          name: sourceFile
          schema:
            type: string
          description: Explicit source filename to remove
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteDocumentInput"
      responses:
        "200":
          description: Document removal result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteDocumentResult"
        "400":
          description: Missing or unsafe selector
        "404":
          description: Document not found

  /rag/documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [rag]
      summary: Remove one ingested document from the candidate index by document ID
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
          description: Explicit document ID to remove
      responses:
        "200":
          description: Document removal result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteDocumentResult"
        "400":
          description: Unsafe selector
        "404":
          description: Document not found

  /rag/chunks:
    get:
      operationId: listChunks
      tags: [rag]
      summary: List all chunks in the index
      parameters:
        - in: query
          name: documentId
          schema:
            type: string
          description: Filter by document ID
      responses:
        "200":
          description: List of chunks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RagChunk"

  /rag/ingest:
    post:
      operationId: ingestDocuments
      tags: [rag]
      summary: Ingest or restore PDFs from the configured source folders
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestInput"
      responses:
        "200":
          description: Ingestion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"

  /rag/reindex:
    post:
      operationId: reindexActiveDocuments
      tags: [rag]
      summary: Rebuild chunks and vectors for currently indexed documents only
      responses:
        "200":
          description: Active-document reindex result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"

  /rag/reset:
    post:
      operationId: resetIndex
      tags: [rag]
      summary: Reset the vector index and all ingested data
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetInput"
      responses:
        "200":
          description: Reset result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetResult"

  /rag/query:
    post:
      operationId: ragQuery
      tags: [rag]
      summary: Query the RAG pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryInput"
      responses:
        "200":
          description: RAG answer with citations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"

  /rag/upload:
    post:
      operationId: uploadDocument
      tags: [rag]
      summary: Upload a PDF file for ingestion
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Upload and ingestion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"

  /rag/reports:
    get:
      operationId: getReports
      tags: [rag]
      summary: Get comparison reports
      responses:
        "200":
          description: Comparison reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComparisonReport"

  /rag/test-model:
    get:
      operationId: testModel
      tags: [rag]
      summary: Test generation model connectivity
      responses:
        "200":
          description: Model connectivity test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelTestResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RagStatus:
      type: object
      required: [documentCount, chunkCount, vectorCount, ready, embeddingModel, generationProvider, generationModel, ollamaBaseUrl]
      properties:
        documentCount:
          type: integer
        chunkCount:
          type: integer
        vectorCount:
          type: integer
        ready:
          type: boolean
        embeddingModel:
          type: string
        generationProvider:
          type: string
          enum: [ollama, openai]
        generationModel:
          type: string
        ollamaBaseUrl:
          type: string
        openaiBaseUrl:
          type: ["string", "null"]
        lastIngestedAt:
          type: ["string", "null"]
        notes:
          type: array
          items:
            type: string

    RagDocument:
      type: object
      required: [id, filename, pageCount, chunkCount, ingestedAt, status]
      properties:
        id:
          type: string
        filename:
          type: string
        displayTitle:
          type: string
          nullable: true
        pageCount:
          type: integer
        chunkCount:
          type: integer
        vectorCount:
          type: integer
          nullable: true
        ingestedAt:
          type: string
        status:
          type: string
        cleaningFlags:
          type: array
          items:
            type: string
        cleanupQuality:
          $ref: "#/components/schemas/CleanupQuality"
          nullable: true

    CleanupQuality:
      type: object
      required:
        - tracked
        - artifactsReducedPct
        - originalTrackedArtifacts
        - remainingTrackedArtifacts
        - artifactsRemoved
        - categoriesTracked
        - byType
      properties:
        tracked:
          type: boolean
        artifactsReducedPct:
          type: ["integer", "null"]
          description: Percentage of tracked artifacts removed from raw text to cleaned text. Null when no tracked raw artifacts exist.
        originalTrackedArtifacts:
          type: integer
        remainingTrackedArtifacts:
          type: integer
        artifactsRemoved:
          type: integer
        categoriesTracked:
          type: array
          items:
            type: string
        byType:
          type: object
          additionalProperties:
            type: object
            required: [raw, clean, removed]
            properties:
              raw:
                type: integer
              clean:
                type: integer
              removed:
                type: integer

    DeleteDocumentInput:
      type: object
      properties:
        documentId:
          type: string
        sourceFile:
          type: string
      description: Provide exactly one explicit documentId or sourceFile.

    DeleteDocumentCounts:
      type: object
      required: [documentCount, chunkCount, vectorCount]
      properties:
        documentCount:
          type: integer
        chunkCount:
          type: integer
        vectorCount:
          type: integer

    DeleteDocumentBeforeCounts:
      allOf:
        - $ref: "#/components/schemas/DeleteDocumentCounts"
        - type: object
          required: [targetChunkCount, targetVectorCount]
          properties:
            targetChunkCount:
              type: integer
            targetVectorCount:
              type: integer

    DeleteDocumentResult:
      type: object
      required: [success, documentId, sourceFile, removedFiles, warnings, before, after]
      properties:
        success:
          type: boolean
        documentId:
          type: string
        sourceFile:
          type: string
        removedFiles:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        before:
          $ref: "#/components/schemas/DeleteDocumentBeforeCounts"
        after:
          $ref: "#/components/schemas/DeleteDocumentCounts"

    RagChunk:
      type: object
      required: [chunkId, documentId, sourceFile, pageStart, pageEnd, text]
      properties:
        chunkId:
          type: string
        documentId:
          type: string
        sourceFile:
          type: string
        pageStart:
          type: integer
        pageEnd:
          type: integer
        sectionPath:
          type: ["string", "null"]
        text:
          type: string
        cleaningFlags:
          type: array
          items:
            type: string
        qualityNotes:
          type: array
          items:
            type: string

    IngestInput:
      type: object
      properties:
        rebuild:
          type: boolean
        documentIds:
          type: array
          items:
            type: string

    ResetInput:
      type: object
      properties:
        confirmation:
          type: string

    IngestResult:
      type: object
      required: [success, documentsIngested, chunksCreated, vectorsCreated, durationMs]
      properties:
        success:
          type: boolean
        documentsIngested:
          type: integer
        chunksCreated:
          type: integer
        vectorsCreated:
          type: integer
        durationMs:
          type: integer
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        cleanupReport:
          type: object
          additionalProperties: true

    ResetResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    QueryInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
        topK:
          type: integer
        includeEvidence:
          type: boolean
        includeDebug:
          type: boolean

    EvidenceSufficiency:
      type: string
      enum: [sufficient, partial, weak, insufficient]
      description: |
        Coarse evidence quality verdict derived from retrieved chunk count,
        source diversity, section quality, and hedging detection.
        - sufficient: ≥3 direct-score chunks from ≥2 sources, no severe hedging
        - partial: ≥2 direct-score chunks OR single source with HQ sections
        - weak: ≥1 direct-score chunk or partial-score chunks present
        - insufficient: no evidence or LLM indicated corpus lacks the information

    CitationValidation:
      type: object
      required: [citedNumbers, validCitations, invalidCitations, uncitedChunkIndices, noisyCitedChunks, allValid, warnings]
      properties:
        citedNumbers:
          type: array
          items:
            type: integer
          description: All citation numbers [N] found in the answer text
        validCitations:
          type: array
          items:
            type: integer
          description: Citation numbers that reference existing, low-noise chunks
        invalidCitations:
          type: array
          items:
            type: integer
          description: Citation numbers that reference non-existent sources
        uncitedChunkIndices:
          type: array
          items:
            type: integer
          description: 1-based indices of retrieved chunks not cited in the answer
        noisyCitedChunks:
          type: array
          items:
            type: integer
          description: Citation numbers whose chunks have noiseScore ≥ 0.5
        allValid:
          type: boolean
          description: True if all cited numbers are valid, non-empty, and non-noisy
        warnings:
          type: array
          items:
            type: string

    QueryResult:
      type: object
      required: [answer, confidence, confidenceReason, evidenceSufficiency, citationValidation, sources, retrievedChunks, warnings, durationMs]
      properties:
        answer:
          type: string
        confidence:
          type: string
          enum: [high, medium, low, insufficient]
        confidenceReason:
          type: string
        evidenceSufficiency:
          $ref: "#/components/schemas/EvidenceSufficiency"
        citationValidation:
          $ref: "#/components/schemas/CitationValidation"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/AnswerSource"
        retrievedChunks:
          type: array
          items:
            $ref: "#/components/schemas/RetrievedChunk"
        warnings:
          type: array
          items:
            type: string
        durationMs:
          type: integer
        debugTrace:
          type: object
          additionalProperties: true
          nullable: true

    AnswerSource:
      type: object
      required: [title, documentId, document_id, document, source, page, pageRange, sourceFile, rawFilename, pageStart, pageEnd, snippet, supportLevel, supportLabel]
      properties:
        title:
          type: string
          description: Friendly source title for client-facing provenance cards.
        documentId:
          type: string
        document_id:
          type: string
          description: Backward-compatible snake_case document identifier alias.
        document:
          type: string
          description: Friendly document title for legacy/raw API clients.
        source:
          type: string
          description: Friendly source title for legacy/raw API clients.
        page:
          type: ["string", "null"]
          description: Formatted page range, for example "p. 4" or "pp. 6-7".
        pageRange:
          type: ["string", "null"]
          description: Formatted page range alias for clients that expect page-range provenance.
        sourceFile:
          type: string
        rawFilename:
          type: string
          description: Raw source filename preserved for debugging/provenance.
        displayTitle:
          type: string
          nullable: true
        pageStart:
          type: integer
        pageEnd:
          type: integer
        sectionPath:
          type: ["string", "null"]
        snippet:
          type: string
        supportLevel:
          type: string
        supportLabel:
          type: string
          description: Backward-compatible support label alias for public QA/source-card validators.
        score:
          type: number

    RetrievedChunk:
      type: object
      required: [chunkId, documentId, sourceFile, pageStart, pageEnd, text, score, retrievalMethod]
      properties:
        chunkId:
          type: string
        documentId:
          type: string
        sourceFile:
          type: string
        pageStart:
          type: integer
        pageEnd:
          type: integer
        sectionPath:
          type: ["string", "null"]
        text:
          type: string
        score:
          type: number
        vectorScore:
          type: number
        bm25Score:
          type: number
        rerankScore:
          type: number
        retrievalMethod:
          type: string
        cleaningFlags:
          type: array
          items:
            type: string

    UploadInput:
      type: object
      required: [filename]
      properties:
        filename:
          type: string
          description: Name of uploaded file (actual file sent as multipart/form-data)

    ComparisonReport:
      type: object
      required: [id, question, candidateAnswer, candidateConfidence, candidateSources, evidenceQuality, classification, createdAt]
      properties:
        id:
          type: string
        question:
          type: string
        candidateAnswer:
          type: string
        candidateConfidence:
          type: string
        candidateSources:
          type: array
          items:
            type: object
            additionalProperties: true
        evidenceQuality:
          type: string
        classification:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    ModelTestResult:
      type: object
      required: [embeddingOk, generationOk, embeddingModel, generationProvider, generationModel, ollamaBaseUrl]
      properties:
        embeddingOk:
          type: boolean
        generationOk:
          type: boolean
        embeddingModel:
          type: string
        generationProvider:
          type: string
          enum: [ollama, openai]
        generationModel:
          type: string
        ollamaBaseUrl:
          type: string
        openaiBaseUrl:
          type: ["string", "null"]
        embeddingError:
          type: ["string", "null"]
        generationError:
          type: ["string", "null"]
        embeddingDims:
          type: ["integer", "null"]
        generationSample:
          type: ["string", "null"]